Add CLI interface and functional subcommand "new"
Added POSIX interface for implementing mkdir, chdir, and filename
splitting; which is problematic as it makes it system-dependent.

Use gfortran chdir, getcwd extensions and a command-based mkdir
instead?

Changed behavior of "new" to allow updating and warn instead of
overwrite existing files. Early on the Haskell version wiped out
some existing README.md and other files; looks like it now stops
if not creating new directory instead?

The build option is failing with the TOML changes. I thought I
tried it and it worked but a new version I checked out fails as well.