Update: hello_complex_2 to expose link bug
There's a bug which causes app-local modules to be added twice if auto-discovery is on and the app is specified in the manifest. This causes the module to be compiled and linked twice. Not detected before because the module contained no symbols. This commit adds an integer symbol to an app-local module to test this bug.