Changed so example/ directory is not created by default
Also changed the sample program to demo.f90 and the test directory to
check.f90 so if they are added with --backfill that reasonable names are
created without modifying the manifest file (fpm.toml) so they can will
still be built with auto-discovery reasonably. Note test/check.f90 was
used instead of test/test.f90 because any program called "test" on ULS
platforms is very problematic.

After having a few users without `fpm` experience try it, they seem to
be looking more for a template to choose and tools to help move existing
projects to fpm (or move fpm to work more with existing projects) than
what "new" currently does. I think this will change as more programs
originate with fpm instead of needing to be moved to it, but instead
of having "new" do everything having templates like some of the samples
already available pulled down, possibly even with an interactive interface
seems like a better long-term direction.