annotated fpm.toml
So in for a pence, in for a pound. On the "new" subcommand if the
"--full" parameter is supplied a fully annotated "fpm.tom" file
is generated much like products like doxygen(1) produce.

This allows a user to compose a complex package build more easily,
without referring to external documentation to compose the manifest.