Updated documentation
Just a few documentation tweaks.

Since the documentation has to change for this PR anyway, cleaned up a few
dusty corners primarily for the "run" subcommand that needed updated to
reflect building now being the default for a run subcommand, for clarity,
and to reflect the auto-build abilitites not available.