put NAME/.gitignore back
So I put the orginal .gitignore back in, and even made it bigger if
"fpm new -full" is used. Here is why ...

Although I preferred only auto-generating .gitignore in build/ initially,
after reviewing the comments I built a list of public github projects
containing fpm.toml files and pulled them and eliminated most trivial
ones that appeared to just be experiments in using fpm by somewhat
arbitrarily removing those with less than one hundred lines of Fortran
source and found the majority of those remaining had a customized (ie. not
the default one built by the "fpm new" subcommand) .gitignore file;
indicating many projects need a .gitignore file in the top directory,
so might as well keep seeding it(?),

I found seven that did not have a .gitignore at all, reinforcing the
need to put one in the build directory itself, as this still does. For
the most part those appeared to be unintentional and looked like several
were inadvertently storing binaries on github as a result, indicating
the change to making a build/.gitignore file is a good idea.

stdlib-fpm does not have one, but did not have a build/ directory on
github, but the change to put one in build/ itself will prevent that being
a problem for anyone that forks or clones it, so that is covered now.

I found a pattern where it looks like a number of large sites are
setting up to be able to build with other systems other than fpm(1),
which by itself does not need a .gitignore file of any complexity;
just perhaps skipping non-core files like logs and fodder and other
user-specific needs it is not too easy to optimize except by exclusion
(that is, exclude everything except the default files used by fpm,
which does not feel like a good direction).

On the other hand, I found two other others using the default github list
for Fortran projects, and saw several other projects essentially making
their own attempt at the same list, but I think a large .gitignore file
is conflating for a new (and probably typical) user so I put back the
simple one by default when "fpm new" is used, but if "fpm new -full"
is used, add the github recommendations.  The github recommendations
do not directly benefit a pure fpm project, but at least so far that is
still a minority of the sample packages I found; so adding it at all or
adding it via --full is a bit kluge, but seemed like it would make handy
information readily available if perhaps not in the most intuitive way,
but would not conflate basic usage.