correct basename(3f) function and subsequently fix --list option
The --list option was incorrectly trimming pathnames when suffix=.false.
was present, and the meaning of the --list option has changed from
originally being an option to display the full pathnames of targets
to displaying a table of matching target basenames. Since the --runner
command can be used to display the full pathnames and the runner command
defaults to an "echo" command the pathnames can still easily be generated,
but an example was added to reflect that.

As the help text needed editing anyway, changed some whitespace to
conform to the requirements of the txt2man(1) utility, to facilitate
easily generating man-pages and HTML versions of the help text.