Avoid infinite loop if command "fpm-" is in path
If a copy of fpm is renamed fpm- it will be called recursively
because plugin-names are searched for in the form of "fpm-*"
because since there is no test if the plugin suffix is blank
the "fpm-" command will be called.

A relatively minor bug; but it takes a one-line change to fix
and is quite confusing to anyone with a "fpm-" command in their
path.