Generalize the function default_variant for accepting C/C++ flags.
The argument flags was made optional and 2 additional optional arguments
were added i.e. c_flags and cxx_flags.
The default values for featture%flags, feature%c_flags and
feature%cxx_flags were set to an empty string. The latter are set if the
respective optional arguments are defined.

flags, c_flags and cxx_flags are set from the callers
default_release_feature and default_debug_feature.
This way there won't be backward compatibility as the flags were already
defined in the callers.

For now, the c_flags and cxx_flags are limited to the fPIC flag
necessary for the shared libraries except for the variant with the gcc
compiler where equivalent c_flags to flags were defined.
More work is needed to complete the set of c_flags and cxx_flags for
other variants.